rename loginSubmit to login
