+++ app/src/App/Action/LoginAction.php
@@ -4,27 +4,48 @@ declare(strict_types=1);
 
 namespace App\Action;
 
+use App\Domain\Service\UsersService;
 use Interop\Http\ServerMiddleware\DelegateInterface;
 use Interop\Http\ServerMiddleware\MiddlewareInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
-use Zend\Diactoros\Response\HtmlResponse;
-use Zend\Expressive\Template\TemplateRendererInterface;
+use Zend\Diactoros\Response\JsonResponse;
+use Zend\Diactoros\Response\RedirectResponse;
+use Zend\Expressive\Router\RouterInterface;
 
 final class LoginAction implements MiddlewareInterface
 {
     /**
-     * @var TemplateRendererInterface
+     * @var UsersService
      */
-    private $template;
+    private $usersService;
 
-    public function __construct(TemplateRendererInterface $template)
-    {
-        $this->template = $template;
+    /**
+     * @var RouterInterface
+     */
+    private $router;
+
+    public function __construct(
+        UsersService $usersService,
+        RouterInterface $router
+    ) {
+        $this->usersService = $usersService;
+        $this->router = $router;
     }
 
     public function process(ServerRequestInterface $request, DelegateInterface $delegate): ResponseInterface
     {
-        return new HtmlResponse($this->template->render('app::login'));
+        $body = $request->getParsedBody();
+
+        if (empty($body['username']) || empty($body['password'])) {
+            return new JsonResponse([], 400);
+        }
+
+        $this->usersService->register(
+            $body['username'],
+            $body['password']
+        );
+
+        return new JsonResponse([]);
     }
 }
